Updated Newsletter files
All files associated with the newsletter for the project. Test.php was to test the connection to the database before adding the code to the newsletter.php file. HTML file is the only one currently working, whereas the newsletter.php has an issue running the php code in the middle of the file. Each are uploaded on the server